requirements.txt
flask
flask_sqlalchemy
Flask-Login
Flask-Bcrypt
pg8000
pandas
openpyxl
reportlab
flask-wtf
wtforms
wtforms-sqlalchemy
waitress
requests
python-dotenv
python-docx
PyPDF2
Flask-Moment
Flask-Mail
Flask-Migrate
Email-validator